osd/PG: store pg_info_t in xattr, purged_snaps separately
Separate the purged_snaps portion of pg_info_t (the one that gets big)
outside of the xattr, but everything else inside it.  This will eventually
let us avoid writing the second bit when that part of the struct doesn't
change, and xattr updates are much faster than xattr + file updates.

Signed-off-by: Sage Weil <sage@inktank.com>