mds: Update backtrace and sentinel for new layout
If a new layout includes a different pool than a file
was originally created with, set the backtrace in the
new pool, and update the backtrace in the old pool to
include a sentinel pointing to the new pool.

This patch includes a pool field within the backtrace object
that gets set to the data/metadata pool where the backtrace
is written by default, but is used as a pointer to a new pool
if its value doesn't match the pool it is contained in.

This requires a new version (4) for the backtrace structure.

Signed-off-by: Sam Lang <sam.lang@inktank.com>